# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API specification for Candels, a OHLC data API platform.",
        "title": "Candles API",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://candles.onrender.com",
    "basePath": "/",
    "paths": {
        "/data": {
            "get": {
                "description": "The endpoint returns the OHLC points for a particular Symbol for  the given time range",
                "produces": [
                    "application/json"
                ],
                "summary": "returns the OHLC points for the given time range",
                "parameters": [
                    {
                        "type": "string",
                        "example": "BTC",
                        "description": "This is the symbol of the OHLC token",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "10344553332",
                        "description": "UNIX time representation of the start time",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "101019283847",
                        "description": "UNIX time representation of the end time",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "page of response",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 5,
                        "description": "Number of OHLC datapoints per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_controller_ohlc_data.GetOHLCResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "The endpoint takes a small CSV file upload and processes it. Max file size is 30MB.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Takes a CSV file upload and processes it",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV file to be processed",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/generate_url": {
            "get": {
                "description": "The endpoint generates a pre-signed URL for the given file name for uploading on S3, It supports huge files",
                "summary": "Generates a pre-signed URL for the given file name for uploading on S3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_controller_ohlc_data.GeneratePresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/status/{filename}": {
            "get": {
                "description": "The endpoint returns the status of the file processing",
                "produces": [
                    "application/json"
                ],
                "summary": "returns the status of the file processing",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"7d2f5f5c-0b1a-4b1e-9c5e-1c2d3e4f5g6h.csv\"",
                        "description": "This is the filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_controller_ohlc_data.ProcessingStatusEntity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_teezzan_candles_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_teezzan_candles_internal_controller_ohlc_data.GeneratePresignedURLResponse": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_teezzan_candles_internal_controller_ohlc_data.GetOHLCResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_teezzan_candles_internal_controller_ohlc_data.OHLC"
                    }
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "github_com_teezzan_candles_internal_controller_ohlc_data.OHLC": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "unix": {
                    "type": "integer"
                }
            }
        },
        "github_com_teezzan_candles_internal_controller_ohlc_data.ProcessingStatusEntity": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/github_com_teezzan_candles_internal_null.String"
                },
                "file_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_teezzan_candles_internal_httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_teezzan_candles_internal_null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}